rename: 카카오 지도에서 사용하는 컴포넌트 하나의 폴더로 이동, 경로 수정
