+++ src/components/kakao/index.ts
@@ -1,1 +1,3 @@
 export { default as KakaoMapContainer } from './Map';
+export { default as Slider } from './Slider';
+export { default as FilterModal } from './FilterModal';
