+++ src/components/result/index.ts
@@ -1,3 +1,2 @@
 export { default as KakaoMapOpenButton } from './KakaoMapOpenButton';
-export { default as FilterModal } from './FilterModal';
 export { default as HouseTooltip } from './HouseTooltip';
